requirements.txt
pandas
numpy
openpyxl
xlsxwriter
pyarrow
fastparquet
matplotlib
seaborn
scipy
statsmodels
humanize
pyyaml

